Add updateHook (#604)
* Add updateHook

A wrapper around sqlite3_update_hook.

For now only as a low-level operation to Database.

To be useful in projects it will probably need some wrapping in the
worker but right now I have no idea yet how that should look.

* Allow removing the updateHook callback

Also release the callback function when the callback is removed or the
database is closed.

Include the previously omitted database name in the callback args as the
sqlite callback does.

---------

Co-authored-by: Erik Soehnel <erik.soehnel@gmail.com>